+++ sdk/resourcemanager/operationsmanagement/armoperationsmanagement/zz_generated_pollers.go
@@ -10,9 +10,8 @@ package armoperationsmanagement
 
 import (
 	"context"
-	"net/http"
-
 	"github.com/Azure/azure-sdk-for-go/sdk/azcore"
+	"net/http"
 )
 
 // SolutionsCreateOrUpdatePoller provides polling facilities until the operation reaches a terminal state.